olx/requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
kombu==5.2.4
prompt-toolkit==3.0.30
pymongo==3.12.3
pytz==2022.2.1
selenium==3.141.0
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
